Move static helper methods to EmailTools class
